コメントにvalidate追加

+++ app/controllers/comments_controller.rb
@@ -5,21 +5,23 @@ class CommentsController < ApplicationController
     @comment = current_user.comments.new(comment_params)
     if @comment.save
       flash[:notice] = "投稿にコメントしました"
-      redirect_to post_path(@comment.post_id)
     else
-      flash.now[:alert] = "コメントに失敗しました"
-      render "show"
+      flash[:alert] = "コメントの投稿に失敗しました"
     end
+    redirect_to post_path(@comment.post_id)
   end
 
   def destroy
-    @post = Post.find(params[:post_id])
-    comment = @post.comments.find(params[:id])
-    comment.destroy
-    flash[:notice] = "コメントを削除しました"
-    redirect_to post_path(@post.id)
+    post = Post.find(params[:post_id])
+    comment = post.comments.find(params[:id])
+    if current_user.id == comment.user.id
+      comment.destroy
+      flash[:notice] = "コメントを削除しました"
+    else
+      flash[:alert] = "エラーが発生しました"
+    end
+    redirect_to post_path(post.id)
   end
-
 
   private
